Move startup app helpers out of daemon main
